Qt Applet: disable session management
The Qt applet has an autostart .desktop file, so it is started at each
login of XDG-based environments; since the default session management
in Qt applications is to save the application as part of the session,
this leads more instances of the Qt applet to run for each new session,
in case the session configuration is to save & restore it.

The solution is to disable the session management for the Qt applet,
so it will not be saved by the session manager.

Closes: #160