Fix handling of add uevents during scanning
Ensure that the enumeration is completed before processing any uevent.
This is to avoid a race where the kernel is still enumerating the devices
and send the uevent while the parent is being authorised.

To not lose these uevents, we store them in a backlog that is processed
at the end of the scanning.