dbus: make large integers (>24bit) work as rule ID when appending a rule
This change prevents a stack corruption with integers larger than 24bit
by making sure the Boolean is of the same size of what GLib uses for a Boolean
(four bytes rather than one).

Appending a rule to the list of rules, i.e. with the parent id of UINT32_MAX-2,
did not work:

$ dbus-send --system --print-reply --dest=org.usbguard1 /org/usbguard1/Policy org.usbguard.Policy1.appendRule string:"allow" uint32:4294967293 boolean:true
Error org.freedesktop.DBus.Error.Failed: Rule set append: rule: Invalid parent ID
$

Debugging revealed that the GVariant is passed correctly, but somehow
the integer was parsed out wrongly.

The only other case where a Boolean is parsed out of a GVariant correctly uses a gboolean.