daemon: always fire the DevicePolicyChanged signal
Before this change, the signal was guarded by a check for whether the
"target" changed.  That is, it fired whenever a device was present,
inserted, or a rule has changed, *except* if a) a rule has matched and
b) it has changed whatever the outcome under the implicit rule was.
That behaviour is arguably much more difficult to explain than just
firing whenever a device is present, inserted, or had a rule changed.
At the same time, such a signal is much more useful for consumers.

More precisely: Now, consumers can listen to that signal and get to know
what happened to a device, i.e. whether it was accepted or rejected.
Before this change, this was difficult at best, if not impossible.
The other PresentChanged signal exists and it tells you whenever a
devices is inserted. But then you wouldn't know whether the target it
carries is the final one or whether a PolicyChange signal will
come to tell what happened with the device. So you needed to invent a
timeout after which you believe that USBGuard won't throw any further
signals and only then act accordingly, e.g. allow or reject a device.

Debian's code search for DevicePolicyChanged reveals three consumers of
this signal: Chromium (probably for ChromeOS), GNOME, and "py3status"
I am not aware of any other

I have not looked into the use of this signal in Chrome OS.

GNOME's built-in USBGuard front-end is currently unable to reliably tell
whether the device that has been inserted will be allowed or blocked, so
they will appreciate this change.

py3status' goal seems to be to "Allow or Reject newly plugged USB
devices using USBGuard" which is arguably much easier if it's known what
will happen to a device. Thus, I reckon that they, too, appreciate this
change.