doc: include defaults in the description
The intention of this change is to make it easier to understand what
USBGuard does by looking at the docs, in particular the Web site.
The defaults have been taken from usbguard-daemon.conf.in. I haven't
noticed any automatic way to generated the (part) of the documentation
from that file, although it might be clever to do in order to lower the
redundancy.