Added a new signal: DevicePolicyApplied to inform about actions on a device
The DevicePolicyChanged signal is guarded by a check for whether the
"target" changed.  That is, it fired whenever a device was present,
inserted, or a rule has changed, *except* if a) a rule has matched and
b) it has changed whatever the outcome under the implicit rule was.
That behaviour is arguably much more difficult to explain than just
firing whenever a device is present, inserted, or had a rule changed.
At the same time, such a signal is much more useful for consumers.

More precisely: Now, consumers can listen to that signal and get to know
what happened to a device, i.e. whether it was accepted or rejected.
Before this change, this was difficult at best, if not impossible.
The other PresentChanged signal exists and it tells you whenever a
devices is inserted. But then you wouldn't know whether the target it
carries is the final one or whether a PolicyChange signal will
come to tell what happened with the device. So you needed to invent a
timeout after which you believe that USBGuard won't throw any further
signals and only then act accordingly, e.g. allow or reject a device.

This is a new signal in order to not break existing applications.