dbus: check whether the client wanted interactive authentication
Some clients do not want to prompt for passwords, e.g. when authorising
keyboards when the screen is locked, prompting for a password doesn't
make sense because the user cannot enter anything.
This change extracts the flags and checks whether the client called with
G_DBUS_CALL_FLAGS_ALLOW_INTERACTIVE_AUTHORIZATION.

The flag is available since 2.46:
https://libsoup.org/gio/GDBusMessage.html#GDBusMessageFlags
Hence, we update the dependency in configure.ac.