ruleset: detect integer overflow of the ID and bail out
The check is semantically correct, because some IDs are reserved. Here,
we exploit the fact that the LastID is defined to be max_int-2, so when
we increment _id_next often enough, we will eventually reach the
reserved LastID. If that ID is reached, we bail out to prevent further
damage.

The daemon logs the exception and keeps running. But it appears
dysfunctional, i.e. does not notice new device let alone authorise any.