RuleSet.cpp: Fix RuleSet::assignID return value regression
Commit 3672e9f84d2ed55994a0df739c4d6d43316c2272
changed the return value of `RuleSet::assignID`
from "return x++" to "return ++x" which broke CI
and skipped value 0.  This commits restores the old
behavior.

This approach does not win prices in purity,
elegance or performance, but performance doesn't matter
in that function and you are invited to claim those prices
for yourself if you can prvide a pull request that humans
understand and that fully passes existing CI...